requirements.txt
py-clob-client>=0.15.0
requests>=2.31.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.1.0
pytest>=7.4.0
pytest-mock>=3.12.0
